Refactor distance to sphere for capsule support.
The generic distance to sphere assumes that the query point is expressed
in a frame that is *not* the sphere's frame. When re-using the sphere logic
for the capsule, this leads to redundant calculations (we end up swapping
1 tranform-transform multiplication and two transform-vector
multiplications for 2 vector additions and one rotation-vector
multiplication.

Furthermore, it allows us to re-use the sphere code even when the query
point projects onto the capsule's spine.